DuplicatePluginError doesn't seem to belong in engine
