+++ seagulls-engine/src/seagulls/app/__init__.py
@@ -1,6 +1,6 @@
 from ._app_interfaces import ISeagullsApplication
 from ._plugin_client import SeagullsEntryPointsPluginsClient, SeagullsEntryPointsPluginSource
-from ._plugin_exceptions import DuplicatePluginError
+from ._plugin_exceptions import InvalidPluginError
 from ._plugin_interfaces import (
     ApplicationType,
     IPluggableSeagullsApplication,
@@ -24,5 +24,5 @@ __all__ = [
     "ISeagullsPluginClient",
     "ISeagullsApplicationPluginRegistrant",
     # Exceptions
-    "DuplicatePluginError",
+    "InvalidPluginError",
 ]
